requirements.txt
cryptography>=42.0
pycryptodome>=3.20
Pillow>=10.0
customtkinter>=5.2
pandas>=2.0
matplotlib>=3.8
fpdf2>=2.7
